@@ -1,24 +1,37 @@
 {
-    "name": "next-netlify-platform-starter",
-    "version": "0.1.0",
+    "name": "calnic",
+    "version": "1.0.0",
+    "type": "module",
     "private": true,
     "scripts": {
         "dev": "next dev",
         "build": "next build",
         "start": "next start",
-        "lint": "next lint"
+        "lint": "eslint",
+        "lint:fix": "eslint --fix"
     },
     "dependencies": {
-        "next": "14.1.3",
-        "react": "18.2.0",
-        "react-dom": "18.2.0"
+        "@mantine/core": "^8.3.6",
+        "@mantine/form": "^8.3.6",
+        "@mantine/hooks": "^8.3.6",
+        "@netlify/blobs": "^10.3.1",
+        "@tabler/icons-react": "^3.35.0",
+        "blobshape": "^1.0.0",
+        "bright": "^1.0.0",
+        "dayjs": "^1.11.19",
+        "markdown-to-jsx": "^7.7.16",
+        "next": "^16.0.1",
+        "react": "19.2.0",
+        "react-big-calendar": "^1.19.4",
+        "react-dom": "19.2.0",
+        "unique-names-generator": "^4.7.1"
     },
     "devDependencies": {
-        "autoprefixer": "^10.4.18",
-        "daisyui": "^4.0.5",
-        "eslint": "8.57.0",
-        "eslint-config-next": "14.1.3",
+        "@tailwindcss/postcss": "^4.1.15",
+        "babel-plugin-react-compiler": "^1.0.0",
+        "eslint": "8.57.1",
+        "eslint-config-next": "16.0.0-beta.0",
         "postcss": "^8.4.36",
-        "tailwindcss": "^3.4.1"
+        "tailwindcss": "^4.1.15"
     }
 }
