@@ -1,7 +1,44 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    reactStrictMode: true,
-    swcMinify: true
+  reactCompiler: true,
+  
+  redirects() {
+    return [
+      {
+        source: '/docs',
+        destination: 'https://docs.netlify.com/frameworks/next-js/overview/',
+        permanent: false,
+      },
+      {
+        source: '/old-blog/:slug',
+        destination: '/classics',
+        permanent: true,
+      },
+      {
+        source: '/github',
+        destination: 'https://github.com/netlify-templates/next-platform-starter',
+        permanent: false,
+      },
+      {
+        source: '/home',
+        destination: '/',
+        permanent: true,
+      },
+    ];
+  },
+  
+  rewrites() {
+    return [
+      {
+        source: '/api/health',
+        destination: '/quotes/random',
+      },
+      {
+        source: '/blog',
+        destination: '/classics',
+      },
+    ];
+  },
 };
 
-module.exports = nextConfig;
+export default nextConfig;
