@@ -1,25 +1,21 @@
+import '@mantine/core/styles.css';
 import '../styles/globals.css';
-import Footer from '../components/Footer.jsx';
-import Header from '../components/Header.jsx';
+import { ColorSchemeScript, MantineProvider } from '@mantine/core';
+import { theme } from './theme';
 
 export const metadata = {
-    title: {
-        template: '%s | Netlify',
-        default: 'Netlify Starter'
-    }
+    title: 'Calnic Medical Clinic',
+    description: 'Changing Healthcare in Calgary One Patient at a Time'
 };
 
 export default function RootLayout({ children }) {
     return (
         <html lang="en">
-            <body className="text-white bg-blue-800">
-                <div className="flex flex-col min-h-screen px-6 bg-grid-pattern sm:px-12">
-                    <div className="flex flex-col w-full max-w-6xl mx-auto grow">
-                        <Header />
-                        <div className="grow">{children}</div>
-                        <Footer />
-                    </div>
-                </div>
+            <head>
+                <ColorSchemeScript />
+            </head>
+            <body>
+                <MantineProvider theme={theme}>{children}</MantineProvider>
             </body>
         </html>
     );
